FA archiver: Numerous fixes for signed/unsigned and size matching
Compiling with clang flags -Wsign-conversion, -Wshorten-64-to-32 and
-Wconversion exposes numerous inconsistencies about word sizes and
sign usage.  Some of these are annoying features of C, but others are
real if subtle bugs.  Mostly the answer is an excessive number of casts
as implemented in this commit.

Turns out that enabling conversion checks with gcc is a right pain.
Rather a lot of changes here, a lot of compiler pedantry.  The general
rule seems to be to avoid small types for general computation and to
use size_t rather more widely than originally hoped.  Also, int64 ->
double conversions all need to be explicit due to possible loss of
precision.

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/python/fa-archiver@79857 e099a375-04f9-0310-9d5f-a741eaff62e1

+++ src/buffer.c
@@ -330,7 +330,7 @@ size_t reader_block_size(struct reader_state *reader)
 
 uint64_t ts_to_microseconds(const struct timespec *ts)
 {
-    return 1000000 * (uint64_t) ts->tv_sec + ts->tv_nsec / 1000;
+    return 1000000 * (uint64_t) ts->tv_sec + (uint64_t) ts->tv_nsec / 1000;
 }
 
 uint64_t get_timestamp(void)
